requirements.txt
Flask
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
Jinja2
SQLAlchemy
WTForms
Werkzeug
sqlalchemy-migrate
sqlparse
requests
git://github.com/Instamojo/instamojo-py#egg=instamojo
